NanoWealthBank/requirements.txt
absl-py==2.1.0
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
arcface==0.0.8
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
deepface==0.0.93
distro==1.9.0
Django==5.1
django-cors-headers==4.6.0
django-otp==1.5.4
django-recaptcha==4.0.0
dlib==19.24.6
docstring_parser==0.16
easydict==1.13
face-recognition==1.3.0
face_recognition_models==0.3.0
facenet-pytorch==2.6.0
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2024.12.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.4
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==3.1.0a0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.12.1
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
insightface==0.7.3
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lz4==4.4.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
mysqlclient==2.2.4
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.20.1
openai==1.60.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.2.0
prettytable==3.15.1
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyotp==2.9.0
pyparsing==3.2.1
PyPDF2==3.0.1
pyreadline3==3.5.4
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
retina-face==0.0.17
rfc3986==1.5.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
shapely==2.0.6
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stringzilla==3.12.2
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
